Update code based on the feedbacks
