--- body.json
@@ -1,103 +1,0 @@
-{
-    "id": "63dd5645469f9d22b7315d73",
-    "name": "Test Case 100",
-    "desc": "Testing signup functionality",
-    "descData": null,
-    "closed": false,
-    "idOrganization": "63442f6f263509019614ec3e",
-    "idEnterprise": null,
-    "pinned": false,
-    "url": "https://trello.com/b/SgrYlTfx/test-case-1",
-    "shortUrl": "https://trello.com/b/SgrYlTfx",
-    "prefs": {
-        "permissionLevel": "private",
-        "hideVotes": false,
-        "voting": "disabled",
-        "comments": "members",
-        "invitations": "members",
-        "selfJoin": true,
-        "cardCovers": true,
-        "cardCounts": false,
-        "isTemplate": false,
-        "cardAging": "regular",
-        "calendarFeedEnabled": false,
-        "hiddenPluginBoardButtons": [],
-        "switcherViews": [
-            {
-                "viewType": "Board",
-                "enabled": true,
-                "_id": "63daf2f8b2eff03f7cb7dfca"
-            },
-            {
-                "viewType": "Table",
-                "enabled": true,
-                "_id": "63daf2f8b2eff03f7cb7dfcb"
-            },
-            {
-                "viewType": "Calendar",
-                "enabled": false,
-                "_id": "63daf2f8b2eff03f7cb7dfcc"
-            },
-            {
-                "viewType": "Dashboard",
-                "enabled": false,
-                "_id": "63daf2f8b2eff03f7cb7dfcd"
-            },
-            {
-                "viewType": "Timeline",
-                "enabled": false,
-                "_id": "63daf2f8b2eff03f7cb7dfce"
-            },
-            {
-                "viewType": "Map",
-                "enabled": false,
-                "_id": "63daf2f8b2eff03f7cb7dfcf"
-            }
-        ],
-        "background": "green",
-        "backgroundColor": "#0079BF",
-        "backgroundImage": null,
-        "backgroundImageScaled": null,
-        "backgroundTile": false,
-        "backgroundBrightness": "dark",
-        "backgroundBottomColor": "#0079BF",
-        "backgroundTopColor": "#0079BF",
-        "canBePublic": true,
-        "canBeEnterprise": true,
-        "canBeOrg": true,
-        "canBePrivate": true,
-        "canInvite": true
-    },
-    "labelNames": {
-        "green": "",
-        "yellow": "",
-        "orange": "",
-        "red": "",
-        "purple": "",
-        "blue": "",
-        "sky": "",
-        "lime": "",
-        "pink": "",
-        "black": "",
-        "green_dark": "",
-        "yellow_dark": "",
-        "orange_dark": "",
-        "red_dark": "",
-        "purple_dark": "",
-        "blue_dark": "",
-        "sky_dark": "",
-        "lime_dark": "",
-        "pink_dark": "",
-        "black_dark": "",
-        "green_light": "",
-        "yellow_light": "",
-        "orange_light": "",
-        "red_light": "",
-        "purple_light": "",
-        "blue_light": "",
-        "sky_light": "",
-        "lime_light": "",
-        "pink_light": "",
-        "black_light": ""
-    }
-}
