+++ src/main/java/com/client/model/Board.java
@@ -1,4 +1,4 @@
-package com.trello.pojo;
+package com.client.model;
 
 import lombok.Data;
