+++ src/main/java/com/client/model/Limits.java
@@ -1,0 +1,4 @@
+package com.client.model;
+
+public class Limits {
+}
