+++ src/main/java/com/client/utils/PropertyUtils.java
@@ -1,4 +1,4 @@
-package com.trello.utils;
+package com.client.utils;
 
 import java.io.BufferedReader;
 import java.io.FileNotFoundException;
