--- src/main/java/com/trello/bodyPayload/Board.java
@@ -1,111 +1,0 @@
-package com.trello.bodyPayload;
-
-public class Board {
-    public static String boardBody(){
-        String payLoad="{\n" +
-                "    \"name\": \"Test Case 10\",\n" +
-                "    \"desc\": \"Testing signup functionality\",\n" +
-                "    \"descData\": null,\n" +
-                "    \"closed\": false,\n" +
-                "    \"idOrganization\": \"63442f6f263509019614ec3e\",\n" +
-                "    \"idEnterprise\": null,\n" +
-                "    \"pinned\": false,\n" +
-                "    \"url\": \"https://trello.com/b/SgrYlTfx/test-case-1\",\n" +
-                "    \"shortUrl\": \"https://trello.com/b/SgrYlTfx\",\n" +
-                "    \"prefs\": {\n" +
-                "        \"permissionLevel\": \"private\",\n" +
-                "        \"hideVotes\": false,\n" +
-                "        \"voting\": \"disabled\",\n" +
-                "        \"comments\": \"members\",\n" +
-                "        \"invitations\": \"members\",\n" +
-                "        \"selfJoin\": true,\n" +
-                "        \"cardCovers\": true,\n" +
-                "        \"isTemplate\": false,\n" +
-                "        \"cardAging\": \"regular\",\n" +
-                "        \"calendarFeedEnabled\": false,\n" +
-                "        \"hiddenPluginBoardButtons\": [],\n" +
-                "        \"switcherViews\": [\n" +
-                "            {\n" +
-                "                \"viewType\": \"Board\",\n" +
-                "                \"enabled\": true,\n" +
-                "                \"_id\": \"63daf2f8b2eff03f7cb7dfca\"\n" +
-                "            },\n" +
-                "            {\n" +
-                "                \"viewType\": \"Table\",\n" +
-                "                \"enabled\": true,\n" +
-                "                \"_id\": \"63daf2f8b2eff03f7cb7dfcb\"\n" +
-                "            },\n" +
-                "            {\n" +
-                "                \"viewType\": \"Calendar\",\n" +
-                "                \"enabled\": false,\n" +
-                "                \"_id\": \"63daf2f8b2eff03f7cb7dfcc\"\n" +
-                "            },\n" +
-                "            {\n" +
-                "                \"viewType\": \"Dashboard\",\n" +
-                "                \"enabled\": false,\n" +
-                "                \"_id\": \"63daf2f8b2eff03f7cb7dfcd\"\n" +
-                "            },\n" +
-                "            {\n" +
-                "                \"viewType\": \"Timeline\",\n" +
-                "                \"enabled\": false,\n" +
-                "                \"_id\": \"63daf2f8b2eff03f7cb7dfce\"\n" +
-                "            },\n" +
-                "            {\n" +
-                "                \"viewType\": \"Map\",\n" +
-                "                \"enabled\": false,\n" +
-                "                \"_id\": \"63daf2f8b2eff03f7cb7dfcf\"\n" +
-                "            }\n" +
-                "        ],\n" +
-                "        \"background\": \"green\",\n" +
-                "        \"backgroundColor\": \"#0079BF\",\n" +
-                "        \"backgroundImage\": null,\n" +
-                "        \"backgroundImageScaled\": null,\n" +
-                "        \"backgroundTile\": false,\n" +
-                "        \"backgroundBrightness\": \"dark\",\n" +
-                "        \"backgroundBottomColor\": \"#0079BF\",\n" +
-                "        \"backgroundTopColor\": \"#0079BF\",\n" +
-                "        \"canBePublic\": true,\n" +
-                "        \"canBeEnterprise\": true,\n" +
-                "        \"canBeOrg\": true,\n" +
-                "        \"canBePrivate\": true,\n" +
-                "        \"canInvite\": true\n" +
-                "    },\n" +
-                "    \"labelNames\": {\n" +
-                "        \"green\": \"\",\n" +
-                "        \"yellow\": \"\",\n" +
-                "        \"orange\": \"\",\n" +
-                "        \"red\": \"\",\n" +
-                "        \"purple\": \"\",\n" +
-                "        \"blue\": \"\",\n" +
-                "        \"sky\": \"\",\n" +
-                "        \"lime\": \"\",\n" +
-                "        \"pink\": \"\",\n" +
-                "        \"black\": \"\",\n" +
-                "        \"green_dark\": \"\",\n" +
-                "        \"yellow_dark\": \"\",\n" +
-                "        \"orange_dark\": \"\",\n" +
-                "        \"red_dark\": \"\",\n" +
-                "        \"purple_dark\": \"\",\n" +
-                "        \"blue_dark\": \"\",\n" +
-                "        \"sky_dark\": \"\",\n" +
-                "        \"lime_dark\": \"\",\n" +
-                "        \"pink_dark\": \"\",\n" +
-                "        \"black_dark\": \"\",\n" +
-                "        \"green_light\": \"\",\n" +
-                "        \"yellow_light\": \"\",\n" +
-                "        \"orange_light\": \"\",\n" +
-                "        \"red_light\": \"\",\n" +
-                "        \"purple_light\": \"\",\n" +
-                "        \"blue_light\": \"\",\n" +
-                "        \"sky_light\": \"\",\n" +
-                "        \"lime_light\": \"\",\n" +
-                "        \"pink_light\": \"\",\n" +
-                "        \"black_light\": \"\"\n" +
-                "    }\n" +
-                "}";
-                return payLoad;
-    }
-
-
-
-}
