--- src/main/java/com/trello/pojo/Limits.java
@@ -1,4 +1,0 @@
-package com.trello.pojo;
-
-public class Limits {
-}
