+++ src/test/java/com/client/tests/BaseTest.java
@@ -1,13 +1,11 @@
-package com.trello.base;
+package com.client.tests;
 
-import com.trello.api.SpecBuilder;
-import com.trello.api.applicationApi.BoardApi;
+import com.client.request.SpecBuilder;
+import com.client.request.BoardApi;
 import io.restassured.specification.RequestSpecification;
 import org.testng.annotations.AfterClass;
 import org.testng.annotations.BeforeClass;
-import org.testng.annotations.Listeners;
 
-@Listeners(com.trello.base.Listeners.class)
 public class BaseTest {
     protected static RequestSpecification requestSpec;
     protected static String boardID;
