--- src/test/java/com/trello/base/Listeners.java
@@ -1,8 +1,0 @@
-package com.trello.base;
-
-import org.testng.ITestListener;
-
-public class Listeners implements ITestListener  {
-
-}
-
