+++ src/test/resources/suites/regression.xml
@@ -3,7 +3,7 @@
 <suite name="RestAssured Automation Tests">
     <test name="Test" parallel="classes">
         <packages>
-            <package name="com.trello.tests"/>
+            <package name="com.client.tests"/>
         </packages>
     </test>
 </suite>
