+++ src/test/resources/suites/smoke.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE suite SYSTEM "https://testng.org/testng-1.0.dtd">
+<suite name="RestAssured Automation Tests">
+    <test name="Test" parallel="classes">
+        <packages>
+            <package name="com.client.tests"/>
+        </packages>
+    </test>
+</suite>
